requirements.txt
livekit-agents[deepgram,openai,cartesia,silero,turn-detector,base]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv
torch
aiohttp>=3.8.0
httpx
numpy
tavily-python